--- a/mathics/algorithm/parts.py
+++ b/mathics/algorithm/parts.py
@@ -6,9 +6,9 @@ Algorithms to access and manipulate elements in nested lists / expressions
 
 
 from mathics.core.expression import Expression
-from mathics.core.symbols import Symbol
+from mathics.core.symbols import Symbol, SymbolList
 from mathics.core.atoms import Integer, from_python
-from mathics.core.systemsymbols import SymbolInfinity
+from mathics.core.systemsymbols import SymbolInfinity, SymbolAll
 from mathics.core.subexpression import SubExpression
 
 from mathics.builtin.exceptions import (
@@ -18,7 +18,9 @@ from mathics.builtin.exceptions import (
     PartRangeError,
 )
 
+
 SymbolNothing = Symbol("Nothing")
+SymbolSpan = Symbol("Span")
 
 
 def join_lists(lists):
@@ -121,7 +123,7 @@ def _parts_span_selector(pspec):
     if len(pspec.leaves) > 1:
         stop = pspec.leaves[1].get_int_value()
         if stop is None:
-            if pspec.leaves[1].get_name() == "System`All":
+            if pspec.leaves[1] is SymbolAll:
                 stop = None
             else:
                 raise MessageException("Part", "span", pspec)
@@ -187,11 +189,11 @@ def _part_selectors(indices):
     the kind of specifications in `indices`.
     """
     for index in indices:
-        if index.has_form("Span", None):
+        if index.has_form(SymbolSpan, None):
             yield _parts_span_selector(index)
         elif index.get_name() == "System`All":
             yield _parts_all_selector()
-        elif index.has_form("List", None):
+        elif index.has_form(SymbolList, None):
             yield _parts_sequence_selector(index.leaves)
         elif isinstance(index, Integer):
             yield _parts_sequence_selector(index), lambda x: x[0]
@@ -354,7 +356,7 @@ def python_levelspec(levelspec):
         else:
             return value
 
-    if levelspec.has_form("List", None):
+    if levelspec.has_form(SymbolList, None):
         values = [value_to_level(leaf) for leaf in levelspec.leaves]
         if len(values) == 1:
             return values[0], values[0]
@@ -447,7 +449,7 @@ def convert_seq(seq):
             stop = value
         else:
             start = value
-    elif seq.has_form("List", 1, 2, 3):
+    elif seq.has_form(SymbolList, 1, 2, 3):
         if len(seq.leaves) == 1:
             start = stop = seq.leaves[0].get_int_value()
             if stop is None: