--- a/mathics/builtin/arithmetic.py
+++ b/mathics/builtin/arithmetic.py
@@ -39,6 +39,10 @@ from mathics.core.symbols import Symbol, SymbolFalse, SymbolList, SymbolTrue
 from mathics.core.systemsymbols import (
     SymbolUndefined,
 )
+
+SymbolSum = Symbol("Sum")
+SymbolProduct = Symbol("Product")
+
 from mathics.core.number import min_prec, dps, SpecialValueError
 
 from mathics.builtin.lists import _IterationFunction
@@ -1163,7 +1167,7 @@ class Sum(_IterationFunction, SympyFunction):
         """
         Perform summation via sympy.summation
         """
-        if expr.has_form("Sum", 2) and expr.leaves[1].has_form("List", 3):
+        if expr.has_form(SymbolSum, 2) and expr.leaves[1].has_form(SymbolList, 3):
             index = expr.leaves[1]
             arg_kwargs = kwargs.copy()
             arg_kwargs["convert_all_global_functions"] = True
@@ -1272,7 +1276,7 @@ class Product(_IterationFunction, SympyFunction):
         return Expression("Times", *items)
 
     def to_sympy(self, expr, **kwargs):
-        if expr.has_form("Product", 2) and expr.leaves[1].has_form("List", 3):
+        if expr.has_form(SymbolProduct, 2) and expr.leaves[1].has_form(SymbolList, 3):
             index = expr.leaves[1]
             try:
                 e_kwargs = kwargs.copy()
@@ -1444,7 +1448,7 @@ class Assuming(Builtin):
         assumptions = assumptions.evaluate(evaluation)
         if assumptions.is_true():
             cond = []
-        elif assumptions.is_symbol() or not assumptions.has_form("List", None):
+        elif assumptions.is_symbol() or not assumptions.has_form(SymbolList, None):
             cond = [assumptions]
         else:
             cond = assumptions._leaves