… CHANGES.rst
… mathics/builtin/assignments/clear.py
… mathics/builtin/assignments/information.py
… mathics/builtin/base.py
… mathics/builtin/box/graphics.py
… mathics/builtin/box/graphics3d.py
… mathics/builtin/box/uniform_polyhedra.py
… mathics/builtin/colors/color_directives.py
… mathics/builtin/colors/color_operations.py
… mathics/builtin/comparison.py
… mathics/builtin/compilation.py
… mathics/builtin/compile/ir.py
… mathics/builtin/drawing/graphics3d.py
… mathics/builtin/drawing/graphics_internals.py
… mathics/builtin/drawing/image.py
… mathics/builtin/drawing/plot.py
… mathics/builtin/files_io/files.py
… mathics/builtin/files_io/importexport.py
… mathics/builtin/functional.py
… mathics/builtin/graphics.py
… mathics/builtin/inference.py
… mathics/builtin/inout.py
… mathics/builtin/intfns/combinatorial.py
… mathics/builtin/intfns/divlike.py
… mathics/builtin/list/associations.py
… mathics/builtin/list/constructing.py
… mathics/builtin/list/eol.py
… mathics/builtin/list/rearrange.py
… mathics/builtin/lists.py
… mathics/builtin/numbers/algebra.py