In this PR, following #47, I propose to modify the behavior of `expression.has_form` to take `Symbol`  references instead of strings as arguments. Again, this avoids having to check the name of the symbol(s) to be compared, and reduces the comparison to a "pointer-like" comparison. 

Benchmarks:


```
from mathics.core.expression import Expression
from mathics.core.symbols import Symbol

SymbolList = Symbol("List")
symboltst = Symbol("System`Internal`MySymboltest")
expr=Expression(symboltst,1,4,Symbol("None"), Symbol("True")) 
expr2=Expression(SymbolList,1,4,Symbol("None"), Symbol("True")) 
```

Old
-----

```
%timeit  expr.has_form("List", None)
```
678 ns ± 4.32 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)


```
%timeit  expr.has_form("System`Internal`MySymboltest", None)
```

1.07 µs ± 6.79 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)


```
%timeit  expr2.has_form("List", None)
```
746 ns ± 6.81 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

```
%timeit  expr2.has_form("System`Internal`MySymboltest", None)
```
996 ns ± 4.15 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)


New
-------
```
%timeit  expr.has_form(SymbolList, None)
```
405 ns ± 2.2 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

```
%timeit  expr.has_form(symboltst, None)
```
449 ns ± 1.84 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

```
%timeit  expr2.has_form(SymbolList, None)
```
462 ns ± 2.18 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```
%timeit  expr2.has_form(symboltst, None)
```
409 ns ± 15.5 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)