@TiagoCavalcante 2021-10-23
> More benchmarks, for dealing with lists:
> 
> ## Old
> ```
> %timeit  expr.has_form(("List", "Rule"), None)
> ```
> 
> 1.05 µs ± 2.74 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> 
> ```
> %timeit  expr.has_form(("System`Internal`MySymboltest","Rule"), None)
> ```
> 
> 1.39 µs ± 2.33 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> 
> ```
> %timeit  expr2.has_form(("List", "Rule"), None)
> ```
> 
> 1.09 µs ± 15.2 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> 
> ```
> %timeit  expr2.has_form(("System`Internal`MySymboltest","Rule"), None)
> ```
> 
> 1.37 µs ± 8.88 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> 
> ## New
> ```
> %timeit  expr.has_form((SymbolList, SymbolRule), None)
> ```
> 
> 715 ns ± 5.14 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> 
> ```
> %timeit  expr.has_form((symboltst,SymbolRule), None)
> ```
> 
> 815 ns ± 7.39 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> 
> ```
> %timeit  expr2.has_form((SymbolList, SymbolRule), None)
> ```
> 
> 811 ns ± 21.4 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> 
> ```
> %timeit  expr2.has_form((symboltst,SymbolRule), None)
> ```
> 
> 712 ns ± 3.7 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

I meant this.