@mmatera 2021-10-24
> >     
> > 
> >     
> >   
> > 712 ns ± 3.7 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> 
> I meant this.

Nop, I didn't edit the output. The units are the right ones.